[Silver II] Title: 나무 자르기, Time: 44 ms, Memory: 2768 KB -BaekjoonHub

+++ 백준/Silver/14247. 나무 자르기/나무 자르기.cc
@@ -5,15 +5,15 @@
 using namespace std;
 
 int main(){
-    long long n;
+    int n;
 
     cin >> n;
-    vector <pair<long long,long long>> grow(n);
-    vector <long long> tree(n);
+    vector <pair<int,int>> grow(n);
+    vector <int> tree(n);
 
-    for (long long i = 0; i < n; i++)
+    for (int i = 0; i < n; i++)
         cin >> tree[i];
-    for (long long i = 0; i < n; i++){
+    for (int i = 0; i < n; i++){
         cin >> grow[i].first;
         grow[i].second = i;
     }
@@ -21,7 +21,7 @@ int main(){
     sort(grow.begin(), grow.end());
 
     long long ans = 0;
-    for (long long i = 0; i < n; i++){
+    for (int i = 0; i < n; i++){
         // cout << grow[i].first << " " << grow[i].second << endl;
         ans += grow[i].first * i + tree[grow[i].second];
     }
